+++ js/consts.js
@@ -1,0 +1,54 @@
+const TITLE = [
+  'заголовок1',
+  'заголовок2',
+  'заголовок3',
+  'заголовок4',
+  'заголовок5',
+  'заголовок6',
+  'заголовок7',
+  'заголовок8',
+  'заголовок9',
+  'заголовок10'
+];
+
+const DESCRIPTION = [
+  'Описание помещения1',
+  'Описание помещения2',
+  'Описание помещения3',
+  'Описание помещения4',
+  'Описание помещения5',
+  'Описание помещения6',
+  'Описание помещения7',
+  'Описание помещения8',
+  'Описание помещения9',
+  'Описание помещения10'
+];
+
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+];
+
+const CHECKTIME = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator'
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+export {TITLE, DESCRIPTION, TYPE, CHECKTIME,FEATURES, PHOTOS};
