+++ js/random-number.js
@@ -1,0 +1,5 @@
+function getRandomNumber(min, max) {
+  const randomInt = Math.random() * (max - min) + min;
+  return Math.floor(randomInt);
+}
+export {getRandomNumber};
